Trying to add cart
unsuccessful due to this error (update-database): Failed executing DbCommand (54ms) [Parameters=[], CommandType='Text', CommandTimeout='30']
ALTER TABLE "Users" ADD CONSTRAINT "FK_Users_Carts_CartId" FOREIGN KEY ("CartId") REFERENCES "Carts" ("Id") ON DELETE CASCADE;
Npgsql.PostgresException (0x80004005): 23503: insert or update on table "Users" violates foreign key constraint "FK_Users_Carts_CartId"

+++ Presentation/ReDoMusic.MVC/Controllers/UsersController.cs
@@ -37,7 +37,7 @@ namespace DoReMusic.MVC.Controllers
                 LastName = lastName,
                 Email = email,
                 Password = password,
-                CreatedOn = DateTime.UtcNow
+                CreatedOn = DateTime.UtcNow,
 
             };
 
@@ -51,8 +51,13 @@ namespace DoReMusic.MVC.Controllers
         {
             var user = _context.Users.Where(x => x.Id == Guid.Parse(id)).FirstOrDefault();
             
-            _context.Users.Remove(user);
-            _context.SaveChanges();
+            if(user != null)
+            {
+                _context.Users.Remove(user);
+                _context.SaveChanges();
+
+            }
+
 
             return RedirectToAction("Index");
         }
